auth_reg/requirements.txt
aiosqlite==0.21.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
cffi==2.0.0
click==8.3.0
colorama==0.4.6
fastapi==0.120.0
h11==0.16.0
idna==3.11
iso8601==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.3
passlib==1.7.4
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
pypika-tortoise==0.6.3
python-multipart==0.0.20
pytz==2025.2
sniffio==1.3.1
starlette==0.48.0
tortoise-orm==0.25.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
